corrected validation & query. working att - report

+++ MainSystem/edit_info_personnel.py
@@ -389,7 +389,7 @@ class EditPersonnelApp:
             and len(personnel_contact_num_var) != 0
             and len(personnel_type_variable) != 0
             and len(personnel_address_var) != 0
-        ):
+            ):
             input_values = [
                 personnel_num_var,
                 personnel_firstname_var,
@@ -401,64 +401,58 @@ class EditPersonnelApp:
                 pesonnel_status_var,
             ]
             concatenated_inputs = "".join(input_values)
-            pattern = re.compile("[^a-zA-Z0-9 \-@.,]")
+            pattern = re.compile("[^a-zA-Z0-9 .,]")
 
             if not pattern.search(concatenated_inputs):
                 if personnel_num_var.isdigit() or (
                     personnel_num_var.startswith("-")
                     and personnel_num_var[1:].isdigit()
-                ):
+                    ):
                     if personnel_contact_num_var.isdigit() or (
                         personnel_contact_num_var.startswith("-")
                         and personnel_contact_num_var[1:].isdigit()
-                    ):
-                        if personnel_firstname_var.replace(" ", "").isalpha():
-                            if personnel_lastname_var.replace(" ", "").isalpha():
-                                if personnel_middlename_var.replace(" ", "").isalpha():
-                                    self.sql_query.update_personnel(
-                                        personnel_num_var,
-                                        personnel_firstname_var,
-                                        personnel_lastname_var,
-                                        personnel_middlename_var,
-                                        personnel_contact_num_var,
-                                        personnel_address_var,
-                                        personnel_type_variable,
-                                        pesonnel_status_var,
-                                    )
-                                    if os.path.exists(self.img_path + "/000000000.jpg"):
-                                        img_name = self.personnel_num
-                                        os.rename(
-                                            self.img_path + "/" + img_name + ".jpg",
-                                            self.img_path + "/000000000.jpg",
-                                        )
-                                    messbx.showinfo(
-                                        "Success",
-                                        "The personnel record has been edited successfully.",
-                                    )
-                                else:
-                                    messbx.showwarning(
-                                        "Warning",
-                                        "The input for the firstname or lastname or middlename is not a valid character.",
-                                    )
-                            else:
-                                messbx.showwarning(
-                                    "Warning",
-                                    "The input for the firstname or lastname or middlename is not a valid character.",
+                        and len(personnel_contact_num_var) == 11
+                        ):
+                        if (personnel_firstname_var.replace(" ", "").isalpha() and 
+                            personnel_lastname_var.replace(" ", "").isalpha() and 
+                            personnel_middlename_var.replace(" ", "").isalpha()
+                            ):
+                            self.sql_query.update_personnel(
+                                personnel_num_var,
+                                personnel_firstname_var,
+                                personnel_lastname_var,
+                                personnel_middlename_var,
+                                personnel_contact_num_var,
+                                personnel_address_var,
+                                personnel_type_variable,
+                                pesonnel_status_var,
+                            )
+                            if os.path.exists(self.img_path + "/000000000.jpg"):
+                                img_name = self.personnel_num
+                                os.rename(
+                                    self.img_path + "/" + img_name + ".jpg",
+                                    self.img_path + "/000000000.jpg",
                                 )
+                            messbx.showinfo(
+                                "Success",
+                                "The personnel's record has been successfully updated.",
+                            )
                         else:
                             messbx.showwarning(
                                 "Warning",
-                                "The input for the firstname or lastname or middlename is not a valid character.",
+                                "There is an invalid character in the input for the name of the personnel.",
                             )
                     else:
                         messbx.showwarning(
                             "Warning",
-                            "The input for the personnel number or contact number is not a valid number.",
+                            "The provided input for the contact number is "+
+                            "invalid and does not correspond to a valid number.",
                         )
                 else:
                     messbx.showwarning(
                         "Warning",
-                        "The input for the personnel number or contact number is not a valid number.",
+                        "The provided input for the personnel number is "+
+                        "invalid and does not correspond to a valid number.",
                     )
             else:
                 messbx.showwarning("Warning", "The input contains special characters.")
