+++ MainSystem/register_visitor.py
@@ -211,7 +211,7 @@ class RegisterVisitorApp:
             and len(first_name_var) != 0
             and len(contact_num_var) != 0
             and len(address_var) != 0
-        ):
+            ):
             input_values = [
                 first_name_var,
                 last_name_var,
@@ -219,48 +219,47 @@ class RegisterVisitorApp:
                 address_var,
             ]
             concatenated_inputs = "".join(input_values)
-            pattern = re.compile("[^a-zA-Z0-9 \-@.,]")
+            pattern = re.compile("[^a-zA-Z0-9 .,]")
 
             if not pattern.search(concatenated_inputs):
-                if contact_num_var.isdigit() or (
-                    contact_num_var.startswith("-") and contact_num_var[1:].isdigit()
-                ):
-                    if first_name_var.replace(" ", "").isalpha():
-                        if last_name_var.replace(" ", "").isalpha():
-                            self.sql_query.register_visitor(
-                                first_name_var,
-                                last_name_var,
-                                contact_num_var,
-                                address_var,
-                            )
-                            img_name = self.sql_query.capture_visitor_image(
-                                first_name_var,
-                                last_name_var,
-                                contact_num_var,
-                                address_var,
-                            )
-                            os.rename(
-                                self.img_path + "/000000000.jpg",
-                                self.img_path + "/" + str(img_name[0]) + ".jpg",
-                            )
-                            messbx.showinfo(
-                                "Success",
-                                "The visitor record has been registered successfully.",
-                            )
-                        else:
-                            messbx.showwarning(
-                                "Warning",
-                                "The input for the firstname or lastname is not a valid character.",
-                            )
+                if  contact_num_var.isdigit() or (
+                    contact_num_var.startswith("-") 
+                    and contact_num_var[1:].isdigit()
+                    and len(contact_num_var) == 11
+                    ):
+                    if (first_name_var.replace(" ", "").isalpha() and 
+                        last_name_var.replace(" ", "").isalpha()
+                        ):
+                        self.sql_query.register_visitor(
+                            first_name_var,
+                            last_name_var,
+                            contact_num_var,
+                            address_var,
+                        )
+                        img_name = self.sql_query.capture_visitor_image(
+                            first_name_var,
+                            last_name_var,
+                            contact_num_var,
+                            address_var,
+                        )
+                        os.rename(
+                            self.img_path + "/000000000.jpg",
+                            self.img_path + "/" + str(img_name[0]) + ".jpg",
+                        )
+                        messbx.showinfo(
+                            "Success",
+                            "The visitor's record has been successfully updated.",
+                        )
                     else:
                         messbx.showwarning(
                             "Warning",
-                            "The input for the firstname or lastname is not a valid character.",
+                            "There is an invalid character in the input for the name the visitor.",
                         )
                 else:
                     messbx.showwarning(
                         "Warning",
-                        "The provided input for the contact number is not a valid numeric value.",
+                        "The provided input for the contact number is "+
+                        "invalid and does not correspond to a valid number.",
                     )
             else:
                 messbx.showwarning("Warning", "The input contains special characters.")
